requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

# Core
torch==2.1.0+cpu
numpy<2

# Misc????
setuptools<81

# API
flask>=2.3
flask-cors>=3.0
gunicorn>=21.2

# TTS + phoneme timings
piper-tts==1.3.0
# allosaurus==1.0.2
bournemouth-forced-aligner 
phonemizer 
soundfile
german-transliterate>=0.1.3
soundfile>=0.12
praatio==6.2.0
tgt==1.5
ffmpeg 
ffmpeg-python

# Whisper STT (CPU)
faster-whisper==1.2.0   
ctranslate2==4.6.0
transformers==4.37.2
huggingface-hub==0.20.3
sentencepiece==0.1.99

# misc
python-dotenv>=1.0
requests>=2.31
tiktoken>=0.11
blobfile>=2.0
openai>=1.3.0
pip-license-checker